requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==3.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
